Don't fail validation if one server out of several has issues
Currently, if a validation server has SSL issues, the SSL exception is
raised immediately. If the thread talking to the server with these
issues encounters that before any other server returns an OK response,
the validation fails prematurely.

If there is more than one server being used, we should not raise this
exception. We should instead continue to process the responses from
other servers, just as we would for a "Connection refused" or
similar. Also, we should log SSL errors (and other connection errors)
with a bit more info.